Address CR feedback
Also enforce max rpc timeout of 20 seconds on all cadence API calls

Move retry related constants and helper methods to its own file.
